APIs/slack.com/1.7.0/python-aiohttp-client/tests/test_views_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.default_error_template import DefaultErrorTemplate
from openapi_server.models.default_success_template import DefaultSuccessTemplate


pytestmark = pytest.mark.asyncio

async def test_views_open(client):
    """Test case for views_open

    
    """
    params = [('trigger_id', 'trigger_id_example'),
                    ('view', 'view_example')]
    headers = { 
        'Accept': 'application/json',
        'token': 'token_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/views.open',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_views_publish(client):
    """Test case for views_publish

    
    """
    params = [('user_id', 'user_id_example'),
                    ('view', 'view_example'),
                    ('hash', 'hash_example')]
    headers = { 
        'Accept': 'application/json',
        'token': 'token_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/views.publish',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_views_push(client):
    """Test case for views_push

    
    """
    params = [('trigger_id', 'trigger_id_example'),
                    ('view', 'view_example')]
    headers = { 
        'Accept': 'application/json',
        'token': 'token_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/views.push',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_views_update(client):
    """Test case for views_update

    
    """
    params = [('view_id', 'view_id_example'),
                    ('external_id', 'external_id_example'),
                    ('view', 'view_example'),
                    ('hash', 'hash_example')]
    headers = { 
        'Accept': 'application/json',
        'token': 'token_example',
        'Authorization': 'Bearer special-key',
    }
    response = await client.request(
        method='GET',
        path='/api/views.update',
        headers=headers,
        params=params,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

